Verbiage Fix
My beginning attempts at getting involved are through studying your site/documentation and trying to submit small yet helpful changes, not to be a nitpick or grammar police. That said, feel free to decline this commit, but I thought these changes might be appreciated.

Lines 16-25:

Original Text:

In the IT world, it is not rare occurrence for people to become concerned with the performance of an application… once they reach the end of the project! That is, once everything has been designed at both the functional and the technology levels. And unless you take everything apart again, performance optimization is not exactly an easy task.
On the other hand, Symfony2 was conceived from the start to be fast and to favor performance. By way of comparison, Symfony2 is about 3 times faster than Version 1.4 or than Zend Framework 1.10, while taking up 2 times less memory.

Edited text:

In the IT world, it is not a rare occurrence for people to become concerned with the performance of an application… once they reach the end of the project! That is, once everything has been designed at both the functional and technological level. Even if you were to take everything apart, performance optimization is no easy task.

On the other hand, Symfony2 was conceived from the start to be fast, with a strong emphasis on performance.  By way of comparison, Symfony2 is about 3 times faster than Symfony Version 1.4 and Zend Framework 1.10, while also taking up 2 times less memory.

Thanks Symfony Team,

Rylan